Setup One-To-One relationship

+++ app/src/main/java/com/androbrain/data/LibraryEntity.kt
@@ -6,5 +6,5 @@ import androidx.room.PrimaryKey
 @Entity
 data class LibraryEntity(
     @PrimaryKey val id: Long,
-    val ownerId: Int,
+    val ownerId: Long,
 )
